+++ app/src/main/java/com/aowen/monolith/data/HeroStatistics.kt
@@ -3,6 +3,7 @@ package com.aowen.monolith.data
 data class HeroStatistics(
     val heroId: Int = 0,
     val name: String = "",
+    val heroName: String = "",
     val winRate: Float = 0f,
     val pickRate: Float = 0f,
 )
@@ -10,7 +11,8 @@ data class HeroStatistics(
 fun HeroStatisticsDto.create(): HeroStatistics {
     return HeroStatistics(
         heroId = this.heroId,
-        name = this.displayName,
+        heroName = this.displayName,
+        name = Hero.entries.find { it.heroName == this.displayName }?.pathName ?: "",
         winRate = this.winRate,
         pickRate = this.pickRate,
     )