+++ app/src/main/java/com/aowen/monolith/ui/theme/Theme.kt
@@ -10,6 +10,7 @@ import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import com.aowen.monolith.data.datastore.Theme
 
 val DarkColorScheme = darkColorScheme(
     surface = NeroBlack,
@@ -55,12 +56,16 @@ val LightColorScheme = lightColorScheme(
 
 @Composable
 fun MonolithTheme(
+    localTheme: Theme = Theme.SYSTEM,
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = false,
     content: @Composable () -> Unit
 ) {
     val colorScheme = when {
+        localTheme != Theme.SYSTEM -> {
+            if (localTheme == Theme.DARK) DarkColorScheme else LightColorScheme
+        }
         dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
             val context = LocalContext.current
             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)